% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{ldb_find_tiles}
\alias{ldb_find_tiles}
\title{Find LiDAR tiles that intersect with given features}
\usage{
ldb_find_tiles(db, x)
}
\arguments{
\item{db}{A database connection}

\item{x}{A spatial object supported by the \code{sf} package (e.g. a
spatial data frame with one or more point, line or polygon features)
The object must have a coordinate reference system defined.}
}
\description{
Find LiDAR tiles that intersect with given features
}
