% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{ldb_is_connected}
\alias{ldb_is_connected}
\title{Check if a database connection is valid and open}
\usage{
ldb_is_connected(db)
}
\arguments{
\item{db}{A database connection}
}
\description{
Check if a database connection is valid and open
}
