% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{ldb_load_tile_metadata}
\alias{ldb_load_tile_metadata}
\title{Import LAS tile metadata into the database}
\usage{
ldb_load_tile_metadata(
  db,
  las,
  filename,
  mapname = NULL,
  provider = "nsw_ss",
  purpose = "general"
)
}
\arguments{
\item{db}{An open database connection. For this function, the user must have
administrator rights to the database.}

\item{las}{A LAS object.}

\item{filename}{Path or filename from which the LAS object was read.}

\item{mapname}{Name of the map sheet (assumed to be 100k topographic map) to
assign to this LAS tile. If \code{NULL} (the default), the map name will be
taken from the filename using the regular expression pattern
\code{'^[A-Za-z]+'}}

\item{provider}{Abbreviation for the LiDAR data provider. This value must
appear in the \code{'providers'} database table. The default is 'nsw_ss'
for NSW Spatial Services.}

\item{purpose}{Character value for the 'purpose' column of the database
table. Usually left as the default (\code{'general'}).}
}
\value{
The integer value of the \code{'id'} field for the newly created
  database record.
}
\description{
This function extracts meta-data values from the given LAS object and
writes them to the database. The values are:
}
