% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{ldb_load_vegheight_raster}
\alias{ldb_load_vegheight_raster}
\title{Create a raster table record for vegetation height derived from a LAS tile}
\usage{
ldb_load_vegheight_raster(
  db,
  raster_url,
  raster_crs = NULL,
  tile_id = NULL,
  protocol = "vsicurl",
  tilew = 256,
  R2P = "C:/Program Files/PostgreSQL/12/bin/raster2pgsql.exe",
  host = NULL,
  dbname = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{db}{An open database connection for a user with administrator rights
to the database.}

\item{raster_url}{A URL giving the file location, e.g. in an accessible S3
bucket.}

\item{raster_crs}{The coordinate reference system of the raster file in a
form that can be understood by the \code{sf::st_crs()} function (e.g. a WKT
string or an integer EPSG code. This is used to infer the name of the
relevant meta-data and raster tables in the database. If \code{NULL}
(default), the function will attempt to retrieve the CRS from the raster
file.}

\item{tile_id}{Integer ID for the corresponding existing record in the
'metadata' table. If \code{NULL} (default), the function will search for a
meta-data record based on the file name of the input raster.}

\item{protocol}{The name of a GDAL-supported virtual file system protocol
to use when accessing the raster file from storage. The default is
\code{'vsicurl'}. See
\href{https://gdal.org/user/virtual_file_systems.html}{GDAL online documentation}
for other options. This will be stored as part of the raster file path in
the database raster table record.}

\item{tilew}{The tile size to use for the raster.
See this \href{https://blog.crunchydata.com/blog/postgis-raster-and-crunchy-bridge}{
Postgis raster example} for more details.}

\item{R2P}{The path to the Postgis command line program \code{raster2pgsql}.
If \code{NULL} or not found, the function will attempt to locate the
program on the client system.}

\item{host}{Host/server on which the database resides. If \code{NULL}
(default), will be determined from the database connection object.}

\item{dbname}{Database name. If \code{NULL} (default), will be determined
from the database connection object.}

\item{password}{User password. If \code{NULL}, this will automatically fall
back to the user's \code{pgpass.conf} file, if one exists, or the
\code{PGPASSWORD} environment variable. On a Windows client, the
\code{pgpass.conf} file in the directory
\code{~/AppData/Roaming/postgresql}.}

\item{user}{User name. If \code{NULL} (default), will be determined from the
database connection object.}
}
\description{
This function associates a single-band raster of vegetation height, usually
derived from a LAS point cloud using the \code{CERMBlidar::get_max_height}
function, with an existing meta-data record for the LAS tile. The database
uses \emph{out-db} storage for raster files, where a raster table record
stores the path to the file location, which will usually be AWS S3-compatible
storage although other types of storage can be used.
}
\details{
If any credentials are required to access the file (e.g. AWS Access Key) it
is assumed that this have been set in the environment from which the function
is being called.
}
