% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{pg_table_exists}
\alias{pg_table_exists}
\title{Check if a table exists in the given database}
\usage{
pg_table_exists(db, tablename)
}
\arguments{
\item{db}{An open database connection.}

\item{tablename}{Name of the table (schema is assumed to be 'lidar').}
}
\value{
\code{TRUE} if the table was found; \code{FALSE} otherwise.
}
\description{
Check if a table exists in the given database
}
