% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{ldb_find_tiles}
\alias{ldb_find_tiles}
\title{Find LiDAR tiles that intersect with given features}
\usage{
ldb_find_tiles(db, x)
}
\arguments{
\item{db}{A database connection}

\item{x}{A spatial object with supported by the \code{sf} package: either an
\code{sf} data frame or a \code{sfc} geometry list. The object must have a
coordinate reference system defined. It can contain one or more point, line
or polygon features.}
}
\value{
A data frame with columns 'meta_table' (name of metadata table),
  'srid' (spatial reference identifier / EPSG code for each matching tile,
  'id' (integer tile id, specific to the particular metadata table),
  'filename' (source file name for the LiDAR data),
  'capture_year' (4-digit year number), 'point_density' (average density
  across the tile expressed as points per square metre).
}
\description{
Queries the metadata tables in the database to identify LiDAR tiles that
intersect with the specified features.
}
